context.mode {scNPF}	R Documentation

Building context mode

Description

a context-specific gene-gene network is constructed from the scRNA-seq data by WGCNA package.

Usage

context.mode(data, qt.gene = 0.4, qt.cell = 0.5, nThreads = 1)

Arguments

data
A expression count matrix. The rows correspond to genes and the columns correspond to cells.

qt.gene
A numeric value between 0 and 1 (default: 0.4) indicating the top percent of expressed genes to be reserved for buliding a context-specific gene-gene network. Used only if network = "context".

qt.cell
A numeric value between 0 and 1 (default: 0.5) indicating the top percent of expressed cells to be reserved for buliding a context-specific gene-gene network. Used only if network = "context".

nThreads
The number of cores to use. Default is 1.

Value

A context-specific gene-gene network.

Examples

#Loading example scRNA-seq data.
load(system.file("data","yan.Rdata",package = "scNPF"))
exp.data <- yan$data
dim(exp.data)
label <- yan$label

#Parallel
context.network <- context.mode(data=exp.data,nThreads=8)
dim(context.network)
class(context.network)

[Package scNPF version 0.1.0 Index]