plotHeatmap {scNPF}	R Documentation

Plot Heatmap

Description

This function plots a heatmap figure form a distance or similarity matrix. The colors denote the sample types based on label.

Usage

plotHeatmap(dists, label)

Arguments

dists
A distance or similarity matrix.

label
True labels of sample.

Value

This function returns a heatmap object invisibly.

Examples

library(gplots)
library(RColorBrewer)
load(system.file("data","yan.Rdata",package = "scNPF"))
data.dist <- as.matrix(dist(t(yan$data)))
#dev.off()
plotHeatmap(data.dist,yan$label)

[Package scNPF version 0.1.0 Index]