scNPF.pro {scNPF}	R Documentation

Imputing dropout values of scRNA-seq data.

Description

scNPF-propagation for imputing dropouts and correcting expression expression measurements.scNPF-propagation involves a network propagation process based on random walk with restart (RWR) on a given gene-gene interaction network to obtain a distribution for each node (gene), which captures its relevance to all other genes in the network.

Usage

scNPF.pro(x, network, gamma = 0.5, qt.gene = 0.4, qt.cell = 0.5,
  nThreads = 1)

Arguments

x
A expression count matrix. The rows correspond to genes and the columns correspond to cells.

network
A adjacency matrix contation gene-gene interaction network. User can use priori mode or context mode. For priori mode, users can use publicly available molecular networks. In this package, we provided three human gene-gene interaction networks, including String, HumanNet and an integrated network. For context mode (default), a context-specific gene-gene network is constructed from the scRNA-seq data by WGCNA package.

qt.gene
A numeric value between 0 and 1 (default: 0.4) indicating the top percent of expressed genes to be reserved for buliding a context-specific gene-gene network. Used only if network = "context".

qt.cell
A numeric value between 0 and 1 (default: 0.5) indicating the top percent of expressed cells to be reserved for buliding a context-specific gene-gene network. Used only if network = "context".

nThreads
The number of cores to use. Default is 1.

gammma
A number between 0 and 1 (default: 0.5). gamma is the trade-off between prior information and network diffusion, governing the distance that a signal is allowed to diffuse through the network during smoothing. The specific value of gamma has little effect on the results of network propagation over a sizable range.

Value

A network-smoothed gene expression matrix.

Examples

#Loading example scRNA-seq data.
load(system.file("data","yan.Rdata",package = "scNPF"))
#Testing with all genes
exp.data <- yan$data
#Or testing with paritial genes
#exp.data <- yan$data[1:2000,]

##For context mode
context.data <- scNPF.pro(x=exp.data, network="context",nThreads=8)
dim(context.data)
dim(exp.data)
context.data[1:5,1:3]
exp.data[1:5,1:3]

##For priori mode
##Using String network
load(system.file("data","string.Rdata",package = "scNPF"))
string.data <- scNPF.pro(x=exp.data, network=string,nThreads=8)

##Using HumanNet network
load(system.file("data","humannet.Rdata",package = "scNPF"))
hm.data <- scNPF.pro(x=exp.data,network=humannet,nThreads=8)

##Using integrated network
load(system.file("data","integrated.Rdata",package = "scNPF"))
inter.data <- scNPF.pro(x=exp.data,network=INet,nThreads=8)

[Package scNPF version 0.1.0 Index]