smoothRWR {scNPF}	R Documentation

Imputing dropout values of scRNA-seq data.

Description

scNPF-propagation for imputing dropouts and correcting expression expression measurements.scNPF-propagation involves a network propagation process based on random walk with restart (RWR) on a given gene-gene interaction network to obtain a distribution for each node (gene), which captures its relevance to all other genes in the network.

Usage

smoothRWR(x, network = "context", gamma = 0.5, nThreads = 1)

Arguments

x
A expression count matrix. The rows correspond to genes and the columns correspond to cells.

network
A adjacency matrix contation gene-gene interaction network.

nThreads
The number of cores to use. Default is 1.

gammma
A number between 0 and 1 (default: 0.5). gamma is the trade-off between prior information and network diffusion, governing the distance that a signal is allowed to diffuse through the network during smoothing. The specific value of gamma has little effect on the results of network propagation over a sizable range.

Value

A network-smoothed gene expression matrix.

[Package scNPF version 0.1.0 Index]